+++ src/dcos_e2e/backends/__init__.py
@@ -3,12 +3,12 @@ Classes to allow backend-specific configuration for cluster backend types.
 """
 
 from ._base_classes import ClusterBackend, ClusterManager
-from ._dcos_docker import DCOS_Docker
+from ._docker import Docker
 from ._existing_cluster import ExistingCluster
 
 __all__ = [
     'ClusterBackend',
     'ClusterManager',
-    'DCOS_Docker',
+    'Docker',
     'ExistingCluster',
 ]