+++ tests/backends/test_docker.py
@@ -1,5 +1,5 @@
 """
-Tests for the DC/OS Docker backend.
+Tests for the Docker backend.
 """
 
 import uuid
@@ -9,7 +9,7 @@ from pathlib import Path
 # are disabled.
 from py.path import local  # pylint: disable=no-name-in-module, import-error
 
-from dcos_e2e.backends import DCOS_Docker
+from dcos_e2e.backends import Docker
 from dcos_e2e.cluster import Cluster
 
 
@@ -36,7 +36,7 @@ class TestCustomMasterMounts:
                 'mode': 'rw',
             },
         }
-        backend = DCOS_Docker(custom_master_mounts=custom_master_mounts)
+        backend = Docker(custom_master_mounts=custom_master_mounts)
 
         with Cluster(
             cluster_backend=backend,