+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 
 - [Changelog](#changelog)
+  - [Next](#next)
   - [2017.11.02.0](#201711020)
   - [2017.10.04.0](#201710040)
   - [2017.08.11.0](#201708110)
@@ -23,6 +24,10 @@
 <!--lint enable list-item-bullet-indent-->
 
 # Changelog
+
+## Next
+
+* Backwards incompatible change: Rename `DCOS_Docker` backend to `Docker` backend.
 
 ## 2017.11.02.0